Monstrous maze

+++ src/challenges/monstrous_maze/monstrous_maze_input.rs
@@ -3,5 +3,5 @@ use serde::{Deserialize, Serialize};
 #[derive(PartialEq, Serialize, Deserialize, Debug)]
 pub struct MonstrousMazeInput {
     pub grid: String,
-    pub endurance: u32,
+    pub endurance: u8,
 }
